+++ provider/gce/environ_firewall.go
@@ -16,7 +16,7 @@ import (
 	"github.com/juju/juju/core/network/firewall"
 	"github.com/juju/juju/environs/context"
 	"github.com/juju/juju/provider/common"
-	"github.com/juju/juju/provider/gce/google"
+	"github.com/juju/juju/provider/gce/internal/google"
 )
 
 // globalFirewallName returns the name to use for the global firewall.
