+++ provider/gce/internal/google/auth.go
