+++ provider/gce/internal/google/auth_test.go
