+++ provider/gce/internal/google/gce.go
