+++ provider/gce/upgrades.go
@@ -9,7 +9,7 @@ import (
 	"github.com/juju/juju/environs"
 	"github.com/juju/juju/environs/context"
 	"github.com/juju/juju/environs/tags"
-	"github.com/juju/juju/provider/gce/google"
+	"github.com/juju/juju/provider/gce/internal/google"
 )
 
 // UpgradeOperations is part of the upgrades.OperationSource interface.
